浙大社团大战

+++ yiyingliService/src/cn/yiyingli/ExchangeData/ExContentAndPage.java
@@ -11,20 +11,20 @@ public class ExContentAndPage {
 		map.put("weight", contentAndPage.getWeight());
 		map.put("contentStyle", contentAndPage.getStyle());
 		switch (contentAndPage.getStyle()) {
-		case ContentAndPageDao.STYLE_PASSAGE:
-			map.put("passageId", contentAndPage.getPassage().getId());
-			map.put("passageTitle", contentAndPage.getPassage().getTitle());
-			break;
-		case ContentAndPageDao.STYLE_SERVICEPRO:
-			map.put("serviceProId", contentAndPage.getServicePro().getId());
-			map.put("serviceProTitle", contentAndPage.getServicePro().getTitle());
-			break;
-		case ContentAndPageDao.STYLE_TEACHER:
-			map.put("teacherId", contentAndPage.getTeacher().getId());
-			map.put("teacherName", contentAndPage.getTeacher().getName());
-			break;
-		default:
-			break;
+			case ContentAndPageDao.STYLE_PASSAGE:
+				map.put("passageId", contentAndPage.getPassage().getId());
+				map.put("passageTitle", contentAndPage.getPassage().getTitle());
+				break;
+			case ContentAndPageDao.STYLE_SERVICEPRO:
+				map.put("serviceProId", contentAndPage.getServicePro().getId());
+				map.put("serviceProTitle", contentAndPage.getServicePro().getTitle());
+				break;
+			case ContentAndPageDao.STYLE_TEACHER:
+				map.put("teacherId", contentAndPage.getTeacher().getId());
+				map.put("teacherName", contentAndPage.getTeacher().getName());
+				break;
+			default:
+				break;
 		}
 	}
 
@@ -42,7 +42,7 @@ public class ExContentAndPage {
 	public static void assembleTeacherForUser(ContentAndPage contentAndPage, SuperMap map) {
 		map.put("activityDes", contentAndPage.getActivityDes());
 		ExTeacher.assembleSimpleForUser(contentAndPage.getTeacher(), map);
-		map.put("mile", contentAndPage.getTeacher().getMile());
+		map.put("mile", (long) contentAndPage.getTeacher().getMile().floatValue());
 	}
 
 }
